Split e2e into tunnel-free and tunnel suites
Separate the real-device tests by transport so they can target different
device sets and run as distinct CI steps:

  - test/e2e            tunnel-free commands (lockdown/usbmux) - every device
  - test/e2e/tunnel     iOS 17+ tunnel commands (RSD/CoreDevice/instruments)

Shared plumbing (binary build, device list, per-device helpers) moves to an
importable test/e2e/harness package so both suites stay thin. real-device.yml
runs the two suites as separate steps; the tunnel step targets
GO_IOS_E2E_TUNNEL_DEVICES (falling back to GO_IOS_E2E_DEVICES) so older,
non-tunnel devices can be added to the full list without breaking it.
